Refactor `rgbmatrix` imports and Matrix class name
