+++ pyproject.toml
@@ -21,7 +21,6 @@ rgbmatrixemulator = "^0.11.3"
 
 [tool.mypy]
 exclude = ["_local_sandbox"]
-mypy_path = "src"
 follow_imports = "normal"
 disallow_any_generics = true
 disallow_subclassing_any = true
